factor out dev client in api server
